Some refactor on git diff and ignore getting commit information failed on migrating pull request review comments
